Finalização dsa classes com Spring Data

+++ src/main/java/br/ufsm/poow2/biblioteca_rest/model/Autor.java
@@ -28,7 +28,7 @@ public class Autor {
             name = "idAutor",
             updatable = false
     )
-    private Integer idAutor;
+    private Long idAutor;
 
     @Column(
             name = "nomeAutor",
